+++ torchmetrics/functional/audio/pit.py
@@ -127,11 +127,11 @@ def pit(
         best_perm of shape [batch]
 
     Example:
-        >>> from torchmetrics.functional.audio import si_sdr
+        >>> from torchmetrics.functional.audio import scale_invariant_signal_distortion_ratio
         >>> # [batch, spk, time]
         >>> preds = torch.tensor([[[-0.0579,  0.3560, -0.9604], [-0.1719,  0.3205,  0.2951]]])
         >>> target = torch.tensor([[[ 1.0958, -0.1648,  0.5228], [-0.4100,  1.1942, -0.5103]]])
-        >>> best_metric, best_perm = pit(preds, target, si_sdr, 'max')
+        >>> best_metric, best_perm = pit(preds, target, scale_invariant_signal_distortion_ratio, 'max')
         >>> best_metric
         tensor([-5.1091])
         >>> best_perm
@@ -189,11 +189,11 @@ def pit_permutate(preds: Tensor, perm: Tensor) -> Tensor:
         Tensor: the permutated version of estimate
 
     Example:
-        >>> from torchmetrics.functional.audio import si_sdr
+        >>> from torchmetrics.functional.audio import scale_invariant_signal_distortion_ratio
         >>> # [batch, spk, time]
         >>> preds = torch.tensor([[[-0.0579,  0.3560, -0.9604], [-0.1719,  0.3205,  0.2951]]])
         >>> target = torch.tensor([[[ 1.0958, -0.1648,  0.5228], [-0.4100,  1.1942, -0.5103]]])
-        >>> best_metric, best_perm = pit(preds, target, si_sdr, 'max')
+        >>> best_metric, best_perm = pit(preds, target, scale_invariant_signal_distortion_ratio, 'max')
         >>> best_metric
         tensor([-5.1091])
         >>> best_perm