+++ torchmetrics/text/wer.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from typing import Any, Callable, List, Optional, Union
+from warnings import warn
 
 import torch
 from torch import Tensor, tensor
@@ -20,7 +21,7 @@ from torchmetrics.functional.text.wer import _wer_compute, _wer_update
 from torchmetrics.metric import Metric
 
 
-class WER(Metric):
+class WordErrorRate(Metric):
     r"""
     Word error rate (WER_) is a common metric of the performance of an automatic speech recognition system.
     This value indicates the percentage of words that were incorrectly predicted.
@@ -57,7 +58,7 @@ class WER(Metric):
     Examples:
         >>> predictions = ["this is the prediction", "there is an other sample"]
         >>> references = ["this is the reference", "there is another one"]
-        >>> metric = WER()
+        >>> metric = WordErrorRate()
         >>> metric(predictions, references)
         tensor(0.5000)
     """
@@ -100,3 +101,29 @@ class WER(Metric):
             Word error rate score
         """
         return _wer_compute(self.errors, self.total)
+
+
+class WER(WordErrorRate):
+    r"""
+    Word error rate (WER_) is a common metric of the performance of an automatic speech recognition system.
+
+    .. deprecated:: v0.7
+        Use :class:`torchmetrics.WordErrorRate`. Will be removed in v0.8.
+
+    Examples:
+        >>> predictions = ["this is the prediction", "there is an other sample"]
+        >>> references = ["this is the reference", "there is another one"]
+        >>> metric = WER()
+        >>> metric(predictions, references)
+        tensor(0.5000)
+    """
+
+    def __init__(
+        self,
+        compute_on_step: bool = True,
+        dist_sync_on_step: bool = False,
+        process_group: Optional[Any] = None,
+        dist_sync_fn: Callable = None,
+    ):
+        warn("`WER` was renamed to `WordErrorRate` in v0.7 and it will be removed in v0.8", DeprecationWarning)
+        super().__init__(compute_on_step, dist_sync_on_step, process_group, dist_sync_fn)